% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMA_helpers.R
\name{b4_A}
\alias{b4_A}
\title{Linear component of CFMA regression}
\usage{
b4_A(bandit)
}
\arguments{
\item{bandit}{A CFMA bandit}
}
\description{
Returns the A matrix of the regression pertaining to a
contextual bandit
}
