% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMA_helpers.R
\name{b4_b}
\alias{b4_b}
\title{Affine component of CFMA regression}
\usage{
b4_b(bandit)
}
\arguments{
\item{bandit}{A CFMA bandit}
}
\description{
Returns the b matrix of the regression pertaining to a
contextual bandit
}
