% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMA_LINUCB_helpers.R
\name{b4_bandit_parameters.LINUCB}
\alias{b4_bandit_parameters.LINUCB}
\title{LINUCB bandit parameters}
\usage{
\method{b4_bandit_parameters}{LINUCB}(bandit)
}
\arguments{
\item{bandit}{A LINUCB bandit}
}
\description{
Returns a list of the LINUCB policy parameters supplied when
instantiating a LINUCB bandit object. Includes :
\itemize{
\item{alpha}{ Exploration parameter of the LINUCB algorithm}
}
}
