% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_helpers.R
\name{b4_bandit_parameters}
\alias{b4_bandit_parameters}
\title{Returns the specific policy parameters of a bandit object}
\usage{
b4_bandit_parameters(bandit)
}
\arguments{
\item{bandit}{A bandit of any class}
}
\value{
A list containing the policy parameters. These change depending on
the specific policy class
}
\description{
Returns the specific policy parameters of a bandit object
}
