% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMA_helpers.R
\name{b4_booleanReward}
\alias{b4_booleanReward}
\title{booleanReward getter}
\usage{
b4_booleanReward(bandit)
}
\arguments{
\item{bandit}{A CFMA bandit}
}
\description{
Returns wheter the visitor rewards are boolean values as
specified when instantiating a CFMA bandit
}
