% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_helpers.R
\name{b4_choices}
\alias{b4_choices}
\title{Returns the choice history of a bandit object}
\usage{
b4_choices(bandit)
}
\arguments{
\item{bandit}{A bandit of any class}
}
\description{
Returns the choice history of a bandit object
}
