% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMA_TS_policy.R
\name{b4_condition_for_thompson_sampling}
\alias{b4_condition_for_thompson_sampling}
\title{Condition for Thompson Sampling}
\usage{
b4_condition_for_thompson_sampling(S, K = ncol(S), alpha = 1, beta = 1)
}
\arguments{
\item{S}{FMA regression parameter}

\item{K}{number of arms}

\item{alpha, beta}{Thompson Sampling regression parameters}
}
\description{
Returns a list containing the next arm to be pulled and the
expected reward
}
