% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_object_evaluation.R
\name{b4_cumulative_regret}
\alias{b4_cumulative_regret}
\title{Returns the cumulative regret of a bandit object}
\usage{
b4_cumulative_regret(bandit, plot = TRUE, ...)
}
\arguments{
\item{bandit}{A bandit of any class}

\item{plot}{Logical. If TRUE, plots the cumulative regret}

\item{...}{Additional plot arguments}
}
\description{
The cumulative regret is computed as the cumulative sum of the simple regret.
See also \code{\link{b4_simple_regret}}. Additional arguments are passed on
to \code{[base]{plot}}
}
