% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMA_EPSG_wrappers.R
\name{b4_EpsilonGreedy}
\alias{b4_EpsilonGreedy}
\title{EpsilonGreedy algorithm}
\usage{
b4_EpsilonGreedy(visitor_reward, epsilon = 0.25, bandit = NULL)
}
\arguments{
\item{visitor_reward}{Dataframe of integer or numeric values}

\item{epsilon}{Numeric value (optional)}
}
\description{
Generate a matrix to save the results (S).
At each iteration play the best arm with a probability of 1-epsilon and
other arm with probability epsilon
}
\examples{
## Generates 10000 numbers from 2 binomial  distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame(cbind(K1,K2) )
#Run epsilon Greedy algorithm
epsilon_greedy_alloc  <- b4_EpsilonGreedy(visitor_reward,epsilon  = 0.25)
}
