% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMA_EXP3_wrappers.R
\name{b4_EXP3}
\alias{b4_EXP3}
\title{EXP3 algorithm}
\usage{
b4_EXP3(visitor_reward, gamma = 0.05, bandit = NULL)
}
\arguments{
\item{visitor_reward}{Dataframe of integer or numeric values}

\item{gamma}{Numeric value (optional)}
}
\description{
Exponential  Weights  for  Exploration  and  Exploitation (EXP3) bandit strategy. Uses a list of weigths which evolve according
to arm's reward. The gamma parameter is a coefficient for balancing between exploitation and exploration.
Generate a matrix to save the results (S).
\itemize{ At each iteration
\item Update weight parameter for each arm
\item Choose randomly an arm according to the distribution of proba
\item Receives a reward in visitor_reward for the arm and associated iteration
\item Updates the results matrix S.
}
}
\examples{
## Generates 1000 numbers from 2 uniform distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame( cbind(K1,K2) )
EXP3_alloc <- b4_EXP3(visitor_reward)
}
