% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_helpers.R
\name{b4_horizon}
\alias{b4_horizon}
\title{Returns the number of observation in the visitor reward of any bandit}
\usage{
b4_horizon(bandit)
}
\arguments{
\item{bandit}{A bandit of any class}
}
\description{
Returns the number of observation in the visitor reward of any bandit
}
