% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_helpers.R
\name{b4_K}
\alias{b4_K}
\title{Returns the number of arms of a bandit object}
\usage{
b4_K(bandit)
}
\arguments{
\item{bandit}{A bandit of any class}
}
\description{
Returns the number of arms of a bandit object
}
