% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMA_object_evaluation.R
\name{b4_mu}
\alias{b4_mu}
\title{Returns the actual FMA reward expectation vector based on visitor reward}
\usage{
b4_mu(bandit)
}
\arguments{
\item{bandit}{A FMA bandit object}
}
\description{
Returns the actual FMA reward expectation vector based on visitor reward
}
