% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMA_object_evaluation.R
\name{b4_mu_hat}
\alias{b4_mu_hat}
\title{Returns the estimated FMA reward expectation vector}
\usage{
b4_mu_hat(bandit)
}
\arguments{
\item{bandit}{A FMA bandit object}
}
\description{
Returns the estimated FMA reward expectation vector
}
