% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMA_helpers.R
\name{b4_n_f}
\alias{b4_n_f}
\title{Contextual dimension}
\usage{
b4_n_f(bandit)
}
\arguments{
\item{bandit}{A CFMA bandit}
}
\description{
Returns the dimension of a CFMA bandit object, i.e. the number
of columns in the dt matrix
}
