% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMA_LINUCB_helpers.R
\name{b4_proba.LINUCB}
\alias{b4_proba.LINUCB}
\title{proba getter method}
\usage{
\method{b4_proba}{LINUCB}(bandit)
}
\arguments{
\item{bandit}{A LINUCB bandit}
}
\description{
Returns the proba history vector pertaining to a LINUCB bandit
instance
}
