% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_helpers.R
\name{b4_proba}
\alias{b4_proba}
\title{Returns the proba slot of a bandit object}
\usage{
b4_proba(bandit)
}
\arguments{
\item{bandit}{A bandit of a class that implements}
}
\value{
A vector whose meaning depends on the bandit class. Not all classes
uses such a slot.
}
\description{
Returns the proba slot of a bandit object
}
