% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_helpers.R
\name{b4_rewards}
\alias{b4_rewards}
\title{Returns the reward history of a bandit object}
\usage{
b4_rewards(bandit)
}
\arguments{
\item{bandit}{A bandit of any class}
}
\description{
Returns the reward history of a bandit object
}
