% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_object_evaluation.R
\name{b4_simple_regret}
\alias{b4_simple_regret}
\title{Simple regret}
\usage{
b4_simple_regret(bandit, plot = TRUE, ...)
}
\arguments{
\item{bandit}{A bandit of any class}

\item{plot}{Logical. If TRUE, plots the simple regret}

\item{...}{Additional plot arguments}
}
\description{
Returns the simple regret of a bandit object. It is computed as,
for each observation of visitor reward, its maximum minus the reward
gotten. Additional arguments are passed on to \code{\link[base]{plot}}
}
