% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMA_helpers.R
\name{b4_th_hat}
\alias{b4_th_hat}
\title{Estimated reward expectation of a CFMA bandit}
\usage{
b4_th_hat(bandit)
}
\arguments{
\item{bandit}{A CFMA bandit}
}
\description{
Returns the theta_hat matrix of a bandit object
}
