% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMA_object_evaluation.R
\name{b4_theta}
\alias{b4_theta}
\title{Returns the actual CFMA reward expectation vector based on visitor reward}
\usage{
b4_theta(bandit)
}
\arguments{
\item{bandit}{A CFMA bandit object}
}
\description{
Returns the actual CFMA reward expectation vector based on visitor reward
}
