% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_helpers.R
\name{b4_time}
\alias{b4_time}
\title{Returns the elapsed time during the iteration of the policy}
\usage{
b4_time(bandit)
}
\arguments{
\item{bandit}{A bandit of any class}
}
\value{
An object of type "proc_time". See also
\code{\link[base]{system.time}}
}
\description{
Returns the elapsed time during the iteration of the policy
}
