% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMA_UCB_wrappers.R
\name{b4_UCB}
\alias{b4_UCB}
\title{UCB algorithm}
\usage{
b4_UCB(visitor_reward, alpha = 1, bandit = NULL)
}
\arguments{
\item{visitor_reward}{Dataframe of integer or numeric values}

\item{alpha}{Numeric value (optional)}
}
\description{
Generate a matrix to save the results (S).
\itemize{ At each iteration
\item Calculates the arm probabilities
\item Choose the arm with the maximum upper bound (with alpha parameter)
\item Receives a reward in visitor_reward for the arm and associated iteration
\item Updates the results matrix S.uire \code{\link{tic}} and \code{\link{toc}} from \code{\link{tictoc}} library
}
}
\examples{
## Generates 10000 numbers from 2 binomial  distributions
set.seed(4434)
K1 <- rbinom(1000, 1, 0.6)
K2 <- rbinom(1000, 1, 0.7)
## Define a dataframe of rewards
visitor_reward <- as.data.frame(cbind(K1,K2) )
#Run UCB algorithm
ucb_alloc  <- b4_UCB(visitor_reward,alpha = 10)
}
