% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_helpers.R
\name{b4_visitor_reward}
\alias{b4_visitor_reward}
\title{Returns the visitor reward matrix of a bandit object}
\usage{
b4_visitor_reward(bandit)
}
\arguments{
\item{bandit}{A bandit of any class}
}
\description{
Returns the visitor reward matrix of a bandit object
}
