% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_structures.R
\name{empty_BASE}
\alias{empty_BASE}
\title{BASE bandit object helper}
\usage{
empty_BASE(K, ..., class)
}
\arguments{
\item{K}{number of arms}

\item{...}{additional slots for class extension}

\item{class}{class extension names}
}
\description{
Initiates a BASE object
}
