% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMA_object_evaluation.R
\name{FMA_simulate}
\alias{FMA_simulate}
\title{Generic for all FMA algorithm methods}
\usage{
FMA_simulate(bandit, visitor_reward)
}
\arguments{
\item{bandit}{A bandit object on which to append new observations}

\item{visitor_reward}{visitor reward matrix. New observations}
}
\description{
Generic for all FMA algorithm methods
}
