% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_object_evaluation.R
\name{hist.BASE}
\alias{hist.BASE}
\title{Choice frequency histogram}
\usage{
\method{hist}{BASE}(x, ...)
}
\arguments{
\item{...}{Additional histogram plot arguments}

\item{bandit}{A bandit of any class}
}
\description{
Plots the frequency histogram of the choice history of a bandit
object.
}
