% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_structures.R
\name{new_BASE}
\alias{new_BASE}
\title{BASE bandit object constructor}
\usage{
new_BASE(
  K,
  H,
  visitor_reward,
  choices,
  rewards,
  time,
  ...,
  class = character()
)
}
\description{
BASE acts like an abstract class and serves as a foundation for
actual implementation. Their purpose is to store policy parameters,
regression variables and iteration history in a list.

@param K number of arms
@param H horizon, aka number of observation
@param visitor_reward matrix
@param choices choice history
@param rewards reward history
@param time elapsed time during policy iterations
@param ... additional slots for class extension
@param class class extension names
}
