% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFMA_LINUCB_helpers.R
\name{print.LINUCB}
\alias{print.LINUCB}
\title{Printing LINUCB bandit}
\usage{
\method{print}{LINUCB}(x, ...)
}
\arguments{
\item{x}{A LINUCB bandit object}

\item{...}{Additional print arguments}
}
\description{
Printing LINUCB bandit
}
