% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASE_structures.R
\name{validate_BASE}
\alias{validate_BASE}
\title{BASE bandit object validator}
\usage{
validate_BASE(bandit)
}
\arguments{
\item{bandit}{A BASE bandit object}
}
\description{
Checks the class value and whether its slot are valid
}
